lte: (fixes #106) IMSI is a uint64_t

+++ src/lte/model/epc-gtpc-header.h
@@ -334,8 +334,8 @@ public:
   virtual void Print (std::ostream &os) const;
   virtual uint32_t GetMessageSize (void) const;
 
-  uint8_t GetImsi () const;
-  void SetImsi (uint8_t imsi);
+  uint64_t GetImsi () const;
+  void SetImsi (uint64_t imsi);
 
   uint32_t GetUliEcgi () const;
   void SetUliEcgi (uint32_t uliEcgi);
